// generated/src/app/api-models/cap-entity.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapEntity {
  capModelId: string;
  capEntityId: string;
  capEntityCode: string;
  seqSpecProvId: number;
  seqVendId: number;
  affProvKey: string;
  description: string;
  provEqPcp: string;
  participationFlag: string;
  contractType: string;
  useThreshold: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqCapVendAddress: number;
}